move the error provide code to its own module
