+++ library/core/src/error/mod.rs
@@ -1,8 +1,13 @@
-#![doc = include_str!("error.md")]
+#![doc = include_str!("../error.md")]
 #![stable(feature = "error_in_core", since = "1.81.0")]
 
 use crate::any::TypeId;
-use crate::fmt::{self, Debug, Display, Formatter};
+use crate::fmt::{Debug, Display};
+
+mod provide;
+
+#[unstable(feature = "error_generic_member_access", issue = "99301")]
+pub use provide::{Request, request_ref, request_value};
 
 /// `Error` is a trait representing the basic expectations for error values,
 /// i.e., values of type `E` in [`Result<T, E>`].
@@ -153,12 +158,52 @@ pub trait Error: Debug + Display {
     /// Used in conjunction with [`Request::provide_value`] and [`Request::provide_ref`] to extract
     /// references to member variables from `dyn Error` trait objects.
     ///
+    /// Every individual error type can `provide` some types that contain programmatic
+    /// information about the error. The set of values `provide`d by a given error type
+    /// can normally change between different versions of that error type's library. However,
+    /// a library defining an error type can always make stronger backwards-compatibility
+    /// promises - for example, a library can declare that an error type always provides a
+    /// [`Location`](core::panic::Location) that provides a relevant source-code location.
+    ///
+    /// # Whether to provide by reference or by value
+    ///
+    /// [`Request::provide_value`] and [`Request::provide_ref`] are two different namespaces.
+    /// Therefore, when providing a type, it needs to be picked whether it will be provided
+    /// by reference or by value.
+    ///
+    /// If a type is provided by value, then a new copy of that type has to be created every
+    /// time it is provided, but if it is provided by reference, then the provided value has
+    /// to be stored somewhere within the error so that the reference can be returned.
+    ///
+    /// Some general rules:
+    ///
+    /// 1. If a type is [Copy], it is conventional to provide it by value.
+    /// 2. If a type is not [Copy] but also not computed at provide time, for example
+    ///    backtrace types that are captured when the error is created, it is conventional
+    ///    to provide it by reference.
+    ///
+    /// Provided types that are not [Copy] and computed at provide time are fairly rare in
+    /// practice. However, when using them, you should be using
+    /// [`Request::would_be_satisfied_by_value_of`] to avoid computing them when they
+    /// are not requested.
+    ///
+    /// # Common uses of `provide`
+    ///
+    /// 1. [`Location`](core::panic::Location), provided by value, to indicate a source-code
+    ///    location relevant to the error. This allows following the [`Error::source`]
+    ///    chain to generate a "logical" backtrace, even in the absence of debug information.
+    /// 2. A backtrace, provided by reference, that contains the backtrace of the error.
+    /// 3. Various exit code types, normally provided by value. For example, an HTTP framework
+    ///    might request an HTTP status on an error, to allow error types to override the HTTP
+    ///    status returned on an error (consult your framework for specific behavior).
+    ///
     /// # Example
     ///
     /// ```rust
     /// #![feature(error_generic_member_access)]
     /// use core::fmt;
-    /// use core::error::{request_ref, Request};
+    /// use core::error::{request_ref, request_value, Request};
+    /// use core::panic::Location;
     ///
     /// #[derive(Debug)]
     /// enum MyLittleTeaPot {
@@ -180,6 +225,7 @@ pub trait Error: Debug + Display {
     /// #[derive(Debug)]
     /// struct Error {
     ///     backtrace: MyBacktrace,
+    ///     location: Location<'static>,
     /// }
     ///
     /// impl fmt::Display for Error {
@@ -191,17 +237,21 @@ pub trait Error: Debug + Display {
     /// impl std::error::Error for Error {
     ///     fn provide<'a>(&'a self, request: &mut Request<'a>) {
     ///         request
-    ///             .provide_ref::<MyBacktrace>(&self.backtrace);
+    ///             .provide_ref::<MyBacktrace>(&self.backtrace)
+    ///             .provide_value::<Location<'_>>(self.location);
     ///     }
     /// }
     ///
     /// fn main() {
     ///     let backtrace = MyBacktrace::new();
-    ///     let error = Error { backtrace };
+    ///     let location = Location::caller();
+    ///     let error = Error { backtrace, location: *location };
     ///     let dyn_error = &error as &dyn std::error::Error;
     ///     let backtrace_ref = request_ref::<MyBacktrace>(dyn_error).unwrap();
+    ///     let location = request_value::<Location<'_>>(dyn_error).unwrap();
     ///
     ///     assert!(core::ptr::eq(&error.backtrace, backtrace_ref));
+    ///     assert_eq!(error.location, location);
     ///     assert!(request_ref::<MyLittleTeaPot>(dyn_error).is_none());
     /// }
     /// ```
@@ -428,643 +478,6 @@ impl dyn Error {
         // dyn-incompatible). Requiring an `Unsize` bound is not backwards compatible.
 
         Source { current: Some(self) }
-    }
-}
-
-/// Requests a value of type `T` from the given `impl Error`.
-///
-/// # Examples
-///
-/// Get a string value from an error.
-///
-/// ```rust
-/// #![feature(error_generic_member_access)]
-/// use std::error::Error;
-/// use core::error::request_value;
-///
-/// fn get_string(err: &impl Error) -> String {
-///     request_value::<String>(err).unwrap()
-/// }
-/// ```
-#[unstable(feature = "error_generic_member_access", issue = "99301")]
-pub fn request_value<'a, T>(err: &'a (impl Error + ?Sized)) -> Option<T>
-where
-    T: 'static,
-{
-    request_by_type_tag::<'a, tags::Value<T>>(err)
-}
-
-/// Requests a reference of type `T` from the given `impl Error`.
-///
-/// # Examples
-///
-/// Get a string reference from an error.
-///
-/// ```rust
-/// #![feature(error_generic_member_access)]
-/// use core::error::Error;
-/// use core::error::request_ref;
-///
-/// fn get_str(err: &impl Error) -> &str {
-///     request_ref::<str>(err).unwrap()
-/// }
-/// ```
-#[unstable(feature = "error_generic_member_access", issue = "99301")]
-pub fn request_ref<'a, T>(err: &'a (impl Error + ?Sized)) -> Option<&'a T>
-where
-    T: 'static + ?Sized,
-{
-    request_by_type_tag::<'a, tags::Ref<tags::MaybeSizedValue<T>>>(err)
-}
-
-/// Request a specific value by tag from the `Error`.
-fn request_by_type_tag<'a, I>(err: &'a (impl Error + ?Sized)) -> Option<I::Reified>
-where
-    I: tags::Type<'a>,
-{
-    let mut tagged = Tagged { tag_id: TypeId::of::<I>(), value: TaggedOption::<'a, I>(None) };
-    err.provide(tagged.as_request());
-    tagged.value.0
-}
-
-///////////////////////////////////////////////////////////////////////////////
-// Request and its methods
-///////////////////////////////////////////////////////////////////////////////
-
-/// `Request` supports generic, type-driven access to data. Its use is currently restricted to the
-/// standard library in cases where trait authors wish to allow trait implementors to share generic
-/// information across trait boundaries. The motivating and prototypical use case is
-/// `core::error::Error` which would otherwise require a method per concrete type (eg.
-/// `std::backtrace::Backtrace` instance that implementors want to expose to users).
-///
-/// # Data flow
-///
-/// To describe the intended data flow for Request objects, let's consider two conceptual users
-/// separated by API boundaries:
-///
-/// * Consumer - the consumer requests objects using a Request instance; eg a crate that offers
-///   fancy `Error`/`Result` reporting to users wants to request a Backtrace from a given `dyn Error`.
-///
-/// * Producer - the producer provides objects when requested via Request; eg. a library with an
-///   an `Error` implementation that automatically captures backtraces at the time instances are
-///   created.
-///
-/// The consumer only needs to know where to submit their request and are expected to handle the
-/// request not being fulfilled by the use of `Option<T>` in the responses offered by the producer.
-///
-/// * A Producer initializes the value of one of its fields of a specific type. (or is otherwise
-///   prepared to generate a value requested). eg, `backtrace::Backtrace` or
-///   `std::backtrace::Backtrace`
-/// * A Consumer requests an object of a specific type (say `std::backtrace::Backtrace`). In the
-///   case of a `dyn Error` trait object (the Producer), there are functions called `request_ref` and
-///   `request_value` to simplify obtaining an `Option<T>` for a given type.
-/// * The Producer, when requested, populates the given Request object which is given as a mutable
-///   reference.
-/// * The Consumer extracts a value or reference to the requested type from the `Request` object
-///   wrapped in an `Option<T>`; in the case of `dyn Error` the aforementioned `request_ref` and `
-///   request_value` methods mean that `dyn Error` users don't have to deal with the `Request` type at
-///   all (but `Error` implementors do). The `None` case of the `Option` suggests only that the
-///   Producer cannot currently offer an instance of the requested type, not it can't or never will.
-///
-/// # Examples
-///
-/// The best way to demonstrate this is using an example implementation of `Error`'s `provide` trait
-/// method:
-///
-/// ```
-/// #![feature(error_generic_member_access)]
-/// use core::fmt;
-/// use core::error::Request;
-/// use core::error::request_ref;
-///
-/// #[derive(Debug)]
-/// enum MyLittleTeaPot {
-///     Empty,
-/// }
-///
-/// #[derive(Debug)]
-/// struct MyBacktrace {
-///     // ...
-/// }
-///
-/// impl MyBacktrace {
-///     fn new() -> MyBacktrace {
-///         // ...
-///         # MyBacktrace {}
-///     }
-/// }
-///
-/// #[derive(Debug)]
-/// struct Error {
-///     backtrace: MyBacktrace,
-/// }
-///
-/// impl fmt::Display for Error {
-///     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-///         write!(f, "Example Error")
-///     }
-/// }
-///
-/// impl std::error::Error for Error {
-///     fn provide<'a>(&'a self, request: &mut Request<'a>) {
-///         request
-///             .provide_ref::<MyBacktrace>(&self.backtrace);
-///     }
-/// }
-///
-/// fn main() {
-///     let backtrace = MyBacktrace::new();
-///     let error = Error { backtrace };
-///     let dyn_error = &error as &dyn std::error::Error;
-///     let backtrace_ref = request_ref::<MyBacktrace>(dyn_error).unwrap();
-///
-///     assert!(core::ptr::eq(&error.backtrace, backtrace_ref));
-///     assert!(request_ref::<MyLittleTeaPot>(dyn_error).is_none());
-/// }
-/// ```
-///
-#[unstable(feature = "error_generic_member_access", issue = "99301")]
-#[repr(transparent)]
-pub struct Request<'a>(Tagged<dyn Erased<'a> + 'a>);
-
-impl<'a> Request<'a> {
-    /// Provides a value or other type with only static lifetimes.
-    ///
-    /// # Examples
-    ///
-    /// Provides an `u8`.
-    ///
-    /// ```rust
-    /// #![feature(error_generic_member_access)]
-    ///
-    /// use core::error::Request;
-    ///
-    /// #[derive(Debug)]
-    /// struct SomeConcreteType { field: u8 }
-    ///
-    /// impl std::fmt::Display for SomeConcreteType {
-    ///     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    ///         write!(f, "{} failed", self.field)
-    ///     }
-    /// }
-    ///
-    /// impl std::error::Error for SomeConcreteType {
-    ///     fn provide<'a>(&'a self, request: &mut Request<'a>) {
-    ///         request.provide_value::<u8>(self.field);
-    ///     }
-    /// }
-    /// ```
-    #[unstable(feature = "error_generic_member_access", issue = "99301")]
-    pub fn provide_value<T>(&mut self, value: T) -> &mut Self
-    where
-        T: 'static,
-    {
-        self.provide::<tags::Value<T>>(value)
-    }
-
-    /// Provides a value or other type with only static lifetimes computed using a closure.
-    ///
-    /// # Examples
-    ///
-    /// Provides a `String` by cloning.
-    ///
-    /// ```rust
-    /// #![feature(error_generic_member_access)]
-    ///
-    /// use core::error::Request;
-    ///
-    /// #[derive(Debug)]
-    /// struct SomeConcreteType { field: String }
-    ///
-    /// impl std::fmt::Display for SomeConcreteType {
-    ///     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    ///         write!(f, "{} failed", self.field)
-    ///     }
-    /// }
-    ///
-    /// impl std::error::Error for SomeConcreteType {
-    ///     fn provide<'a>(&'a self, request: &mut Request<'a>) {
-    ///         request.provide_value_with::<String>(|| self.field.clone());
-    ///     }
-    /// }
-    /// ```
-    #[unstable(feature = "error_generic_member_access", issue = "99301")]
-    pub fn provide_value_with<T>(&mut self, fulfil: impl FnOnce() -> T) -> &mut Self
-    where
-        T: 'static,
-    {
-        self.provide_with::<tags::Value<T>>(fulfil)
-    }
-
-    /// Provides a reference. The referee type must be bounded by `'static`,
-    /// but may be unsized.
-    ///
-    /// # Examples
-    ///
-    /// Provides a reference to a field as a `&str`.
-    ///
-    /// ```rust
-    /// #![feature(error_generic_member_access)]
-    ///
-    /// use core::error::Request;
-    ///
-    /// #[derive(Debug)]
-    /// struct SomeConcreteType { field: String }
-    ///
-    /// impl std::fmt::Display for SomeConcreteType {
-    ///     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    ///         write!(f, "{} failed", self.field)
-    ///     }
-    /// }
-    ///
-    /// impl std::error::Error for SomeConcreteType {
-    ///     fn provide<'a>(&'a self, request: &mut Request<'a>) {
-    ///         request.provide_ref::<str>(&self.field);
-    ///     }
-    /// }
-    /// ```
-    #[unstable(feature = "error_generic_member_access", issue = "99301")]
-    pub fn provide_ref<T: ?Sized + 'static>(&mut self, value: &'a T) -> &mut Self {
-        self.provide::<tags::Ref<tags::MaybeSizedValue<T>>>(value)
-    }
-
-    /// Provides a reference computed using a closure. The referee type
-    /// must be bounded by `'static`, but may be unsized.
-    ///
-    /// # Examples
-    ///
-    /// Provides a reference to a field as a `&str`.
-    ///
-    /// ```rust
-    /// #![feature(error_generic_member_access)]
-    ///
-    /// use core::error::Request;
-    ///
-    /// #[derive(Debug)]
-    /// struct SomeConcreteType { business: String, party: String }
-    /// fn today_is_a_weekday() -> bool { true }
-    ///
-    /// impl std::fmt::Display for SomeConcreteType {
-    ///     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    ///         write!(f, "{} failed", self.business)
-    ///     }
-    /// }
-    ///
-    /// impl std::error::Error for SomeConcreteType {
-    ///     fn provide<'a>(&'a self, request: &mut Request<'a>) {
-    ///         request.provide_ref_with::<str>(|| {
-    ///             if today_is_a_weekday() {
-    ///                 &self.business
-    ///             } else {
-    ///                 &self.party
-    ///             }
-    ///         });
-    ///     }
-    /// }
-    /// ```
-    #[unstable(feature = "error_generic_member_access", issue = "99301")]
-    pub fn provide_ref_with<T: ?Sized + 'static>(
-        &mut self,
-        fulfil: impl FnOnce() -> &'a T,
-    ) -> &mut Self {
-        self.provide_with::<tags::Ref<tags::MaybeSizedValue<T>>>(fulfil)
-    }
-
-    /// Provides a value with the given `Type` tag.
-    fn provide<I>(&mut self, value: I::Reified) -> &mut Self
-    where
-        I: tags::Type<'a>,
-    {
-        if let Some(res @ TaggedOption(None)) = self.0.downcast_mut::<I>() {
-            res.0 = Some(value);
-        }
-        self
-    }
-
-    /// Provides a value with the given `Type` tag, using a closure to prevent unnecessary work.
-    fn provide_with<I>(&mut self, fulfil: impl FnOnce() -> I::Reified) -> &mut Self
-    where
-        I: tags::Type<'a>,
-    {
-        if let Some(res @ TaggedOption(None)) = self.0.downcast_mut::<I>() {
-            res.0 = Some(fulfil());
-        }
-        self
-    }
-
-    /// Checks if the `Request` would be satisfied if provided with a
-    /// value of the specified type. If the type does not match or has
-    /// already been provided, returns false.
-    ///
-    /// # Examples
-    ///
-    /// Checks if a `u8` still needs to be provided and then provides
-    /// it.
-    ///
-    /// ```rust
-    /// #![feature(error_generic_member_access)]
-    ///
-    /// use core::error::Request;
-    /// use core::error::request_value;
-    ///
-    /// #[derive(Debug)]
-    /// struct Parent(Option<u8>);
-    ///
-    /// impl std::fmt::Display for Parent {
-    ///     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    ///         write!(f, "a parent failed")
-    ///     }
-    /// }
-    ///
-    /// impl std::error::Error for Parent {
-    ///     fn provide<'a>(&'a self, request: &mut Request<'a>) {
-    ///         if let Some(v) = self.0 {
-    ///             request.provide_value::<u8>(v);
-    ///         }
-    ///     }
-    /// }
-    ///
-    /// #[derive(Debug)]
-    /// struct Child {
-    ///     parent: Parent,
-    /// }
-    ///
-    /// impl Child {
-    ///     // Pretend that this takes a lot of resources to evaluate.
-    ///     fn an_expensive_computation(&self) -> Option<u8> {
-    ///         Some(99)
-    ///     }
-    /// }
-    ///
-    /// impl std::fmt::Display for Child {
-    ///     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    ///         write!(f, "child failed: \n  because of parent: {}", self.parent)
-    ///     }
-    /// }
-    ///
-    /// impl std::error::Error for Child {
-    ///     fn provide<'a>(&'a self, request: &mut Request<'a>) {
-    ///         // In general, we don't know if this call will provide
-    ///         // an `u8` value or not...
-    ///         self.parent.provide(request);
-    ///
-    ///         // ...so we check to see if the `u8` is needed before
-    ///         // we run our expensive computation.
-    ///         if request.would_be_satisfied_by_value_of::<u8>() {
-    ///             if let Some(v) = self.an_expensive_computation() {
-    ///                 request.provide_value::<u8>(v);
-    ///             }
-    ///         }
-    ///
-    ///         // The request will be satisfied now, regardless of if
-    ///         // the parent provided the value or we did.
-    ///         assert!(!request.would_be_satisfied_by_value_of::<u8>());
-    ///     }
-    /// }
-    ///
-    /// let parent = Parent(Some(42));
-    /// let child = Child { parent };
-    /// assert_eq!(Some(42), request_value::<u8>(&child));
-    ///
-    /// let parent = Parent(None);
-    /// let child = Child { parent };
-    /// assert_eq!(Some(99), request_value::<u8>(&child));
-    ///
-    /// ```
-    #[unstable(feature = "error_generic_member_access", issue = "99301")]
-    pub fn would_be_satisfied_by_value_of<T>(&self) -> bool
-    where
-        T: 'static,
-    {
-        self.would_be_satisfied_by::<tags::Value<T>>()
-    }
-
-    /// Checks if the `Request` would be satisfied if provided with a
-    /// reference to a value of the specified type.
-    ///
-    /// If the type does not match or has already been provided, returns false.
-    ///
-    /// # Examples
-    ///
-    /// Checks if a `&str` still needs to be provided and then provides
-    /// it.
-    ///
-    /// ```rust
-    /// #![feature(error_generic_member_access)]
-    ///
-    /// use core::error::Request;
-    /// use core::error::request_ref;
-    ///
-    /// #[derive(Debug)]
-    /// struct Parent(Option<String>);
-    ///
-    /// impl std::fmt::Display for Parent {
-    ///     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    ///         write!(f, "a parent failed")
-    ///     }
-    /// }
-    ///
-    /// impl std::error::Error for Parent {
-    ///     fn provide<'a>(&'a self, request: &mut Request<'a>) {
-    ///         if let Some(v) = &self.0 {
-    ///             request.provide_ref::<str>(v);
-    ///         }
-    ///     }
-    /// }
-    ///
-    /// #[derive(Debug)]
-    /// struct Child {
-    ///     parent: Parent,
-    ///     name: String,
-    /// }
-    ///
-    /// impl Child {
-    ///     // Pretend that this takes a lot of resources to evaluate.
-    ///     fn an_expensive_computation(&self) -> Option<&str> {
-    ///         Some(&self.name)
-    ///     }
-    /// }
-    ///
-    /// impl std::fmt::Display for Child {
-    ///     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    ///         write!(f, "{} failed: \n  {}", self.name, self.parent)
-    ///     }
-    /// }
-    ///
-    /// impl std::error::Error for Child {
-    ///     fn provide<'a>(&'a self, request: &mut Request<'a>) {
-    ///         // In general, we don't know if this call will provide
-    ///         // a `str` reference or not...
-    ///         self.parent.provide(request);
-    ///
-    ///         // ...so we check to see if the `&str` is needed before
-    ///         // we run our expensive computation.
-    ///         if request.would_be_satisfied_by_ref_of::<str>() {
-    ///             if let Some(v) = self.an_expensive_computation() {
-    ///                 request.provide_ref::<str>(v);
-    ///             }
-    ///         }
-    ///
-    ///         // The request will be satisfied now, regardless of if
-    ///         // the parent provided the reference or we did.
-    ///         assert!(!request.would_be_satisfied_by_ref_of::<str>());
-    ///     }
-    /// }
-    ///
-    /// let parent = Parent(Some("parent".into()));
-    /// let child = Child { parent, name: "child".into() };
-    /// assert_eq!(Some("parent"), request_ref::<str>(&child));
-    ///
-    /// let parent = Parent(None);
-    /// let child = Child { parent, name: "child".into() };
-    /// assert_eq!(Some("child"), request_ref::<str>(&child));
-    /// ```
-    #[unstable(feature = "error_generic_member_access", issue = "99301")]
-    pub fn would_be_satisfied_by_ref_of<T>(&self) -> bool
-    where
-        T: ?Sized + 'static,
-    {
-        self.would_be_satisfied_by::<tags::Ref<tags::MaybeSizedValue<T>>>()
-    }
-
-    fn would_be_satisfied_by<I>(&self) -> bool
-    where
-        I: tags::Type<'a>,
-    {
-        matches!(self.0.downcast::<I>(), Some(TaggedOption(None)))
-    }
-}
-
-#[unstable(feature = "error_generic_member_access", issue = "99301")]
-impl<'a> Debug for Request<'a> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        f.debug_struct("Request").finish_non_exhaustive()
-    }
-}
-
-///////////////////////////////////////////////////////////////////////////////
-// Type tags
-///////////////////////////////////////////////////////////////////////////////
-
-pub(crate) mod tags {
-    //! Type tags are used to identify a type using a separate value. This module includes type tags
-    //! for some very common types.
-    //!
-    //! Currently type tags are not exposed to the user. But in the future, if you want to use the
-    //! Request API with more complex types (typically those including lifetime parameters), you
-    //! will need to write your own tags.
-
-    use crate::marker::PhantomData;
-
-    /// This trait is implemented by specific tag types in order to allow
-    /// describing a type which can be requested for a given lifetime `'a`.
-    ///
-    /// A few example implementations for type-driven tags can be found in this
-    /// module, although crates may also implement their own tags for more
-    /// complex types with internal lifetimes.
-    pub(crate) trait Type<'a>: Sized + 'static {
-        /// The type of values which may be tagged by this tag for the given
-        /// lifetime.
-        type Reified: 'a;
-    }
-
-    /// Similar to the [`Type`] trait, but represents a type which may be unsized (i.e., has a
-    /// `?Sized` bound). E.g., `str`.
-    pub(crate) trait MaybeSizedType<'a>: Sized + 'static {
-        type Reified: 'a + ?Sized;
-    }
-
-    impl<'a, T: Type<'a>> MaybeSizedType<'a> for T {
-        type Reified = T::Reified;
-    }
-
-    /// Type-based tag for types bounded by `'static`, i.e., with no borrowed elements.
-    #[derive(Debug)]
-    pub(crate) struct Value<T: 'static>(PhantomData<T>);
-
-    impl<'a, T: 'static> Type<'a> for Value<T> {
-        type Reified = T;
-    }
-
-    /// Type-based tag similar to [`Value`] but which may be unsized (i.e., has a `?Sized` bound).
-    #[derive(Debug)]
-    pub(crate) struct MaybeSizedValue<T: ?Sized + 'static>(PhantomData<T>);
-
-    impl<'a, T: ?Sized + 'static> MaybeSizedType<'a> for MaybeSizedValue<T> {
-        type Reified = T;
-    }
-
-    /// Type-based tag for reference types (`&'a T`, where T is represented by
-    /// `<I as MaybeSizedType<'a>>::Reified`.
-    #[derive(Debug)]
-    pub(crate) struct Ref<I>(PhantomData<I>);
-
-    impl<'a, I: MaybeSizedType<'a>> Type<'a> for Ref<I> {
-        type Reified = &'a I::Reified;
-    }
-}
-
-/// An `Option` with a type tag `I`.
-///
-/// Since this struct implements `Erased`, the type can be erased to make a dynamically typed
-/// option. The type can be checked dynamically using `Tagged::tag_id` and since this is statically
-/// checked for the concrete type, there is some degree of type safety.
-#[repr(transparent)]
-pub(crate) struct TaggedOption<'a, I: tags::Type<'a>>(pub Option<I::Reified>);
-
-impl<'a, I: tags::Type<'a>> Tagged<TaggedOption<'a, I>> {
-    pub(crate) fn as_request(&mut self) -> &mut Request<'a> {
-        let erased = self as &mut Tagged<dyn Erased<'a> + 'a>;
-        // SAFETY: transmuting `&mut Tagged<dyn Erased<'a> + 'a>` to `&mut Request<'a>` is safe since
-        // `Request` is repr(transparent).
-        unsafe { &mut *(erased as *mut Tagged<dyn Erased<'a>> as *mut Request<'a>) }
-    }
-}
-
-/// Represents a type-erased but identifiable object.
-///
-/// This trait is exclusively implemented by the `TaggedOption` type.
-unsafe trait Erased<'a>: 'a {}
-
-unsafe impl<'a, I: tags::Type<'a>> Erased<'a> for TaggedOption<'a, I> {}
-
-struct Tagged<E: ?Sized> {
-    tag_id: TypeId,
-    value: E,
-}
-
-impl<'a> Tagged<dyn Erased<'a> + 'a> {
-    /// Returns some reference to the dynamic value if it is tagged with `I`,
-    /// or `None` otherwise.
-    #[inline]
-    fn downcast<I>(&self) -> Option<&TaggedOption<'a, I>>
-    where
-        I: tags::Type<'a>,
-    {
-        if self.tag_id == TypeId::of::<I>() {
-            // SAFETY: Just checked whether we're pointing to an I.
-            Some(&unsafe { &*(self as *const Self).cast::<Tagged<TaggedOption<'a, I>>>() }.value)
-        } else {
-            None
-        }
-    }
-
-    /// Returns some mutable reference to the dynamic value if it is tagged with `I`,
-    /// or `None` otherwise.
-    #[inline]
-    fn downcast_mut<I>(&mut self) -> Option<&mut TaggedOption<'a, I>>
-    where
-        I: tags::Type<'a>,
-    {
-        if self.tag_id == TypeId::of::<I>() {
-            Some(
-                // SAFETY: Just checked whether we're pointing to an I.
-                &mut unsafe { &mut *(self as *mut Self).cast::<Tagged<TaggedOption<'a, I>>>() }
-                    .value,
-            )
-        } else {
-            None
-        }
     }
 }
